requirements.txt
xbee==2.3.1
pyserial
pyyaml
